High-resolution motor commands and drive modes

+++ firmware/robot_2011/cpu/radio_protocol.h
@@ -3,9 +3,9 @@
 #include <stdint.h>
 
 // Speeds given in the most recent forward packet
-extern int_fast8_t wheel_command[4];
-extern int_fast8_t dribble_command;
-extern uint_fast8_t kick_command;
+extern int wheel_command[4];	// -511 to 511
+extern int dribble_command;		// -511 to 511
+extern int kick_command;		// 0 to 255
 
 extern uint32_t rx_lost_time;
